django-starter-shopify-app/requirements.txt
Django==1.9
django-shopify-auth==0.5.0
psycopg2==2.7.1
python-dotenv==0.6.4

